[jk] Style data table scrollbar
