+++ mage_ai/frontend/oracle/styles/mixins.ts
@@ -48,16 +48,3 @@ export const outline = (props, opts: any = {}) => {
     }
   `;
 };
-
-export function hideScrollBar() {
-  return `
-    // for Internet Explorer, Edge
-    -ms-overflow-style: none;
-    // for Firefox
-    scrollbar-width: none;
-    // for Chrome, Safari, and Opera
-    ::-webkit-scrollbar {
-      display: none;
-    }
-  `;
-}
